fix(files): purge derived assets when deleting a MediaAsset
Two more instances of the FK hazard fixed for channel disconnect in #1737.
A thumbnail is its own MediaAsset tied to the source only by
MediaAssetVersion.derivedFromVersionId — a self-FK with NO ACTION on delete —
so any bare delete of a thumbnailed asset raises 23503.

- cleanup-service `deleteEntityFiles` hard-deleted the asset row directly. It
  now drops the thumbnails' S3 objects through the `cleanupAssetThumbnails`
  already exported from that file, then purges the closure.
- media-asset-service `deleteVersion` called `deleteThumbnailsForSource`, which
  looks like it handles thumbnails but only SOFT-deletes the rows, so the hard
  delete of the source version was still blocked. Latent rather than live: the
  function refuses to delete a current version, and every thumbnailed version
  in dev is its asset's current one.

Both go through a new `purgeMediaAssets` helper that expands the derived
closure and deletes it in ONE statement (NO ACTION is checked at statement end,
when every referencing row is already gone), marking the closure's
StorageLocations so nothing is silently orphaned in S3. The inline copy added
to storageCleanupJob in #1737 now uses it too.

Verified A/B against dev, rolled back. Bare delete of a thumbnailed asset:
23503. Same asset through purgeMediaAssets: UPDATE 2, DELETE 2. deleteVersion
reproduced on a synthetic non-current thumbnailed version: old path errors,
new path succeeds.
